inference/requirements-inference.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pillow==10.4.0
mlflow==2.12.1
boto3==1.42.71
torch==2.9.1
torchvision==0.24.1
numpy==1.26.4
pydantic==2.9.0